requirements.txt
crawl4ai==0.4.247
python-dotenv==1.0.1
pydantic==2.10.6